allow specifying filename for uri source
and, i guess, forget about handling Content-Disposition.

the idea behind both is that a download url might have either:

1. a basename that differs from the intended download filename
2. a query string that makes a mess of everything

and you might say 2. is just a specific instance of 1.

usually (but not always) the server sends a Content-Disposition header
with the intended attachment name in these cases, so that would be
one way to handle it, but it seems somehow cleaner to provide an
explicit 'filename' parameter to communicate exactly what was intended