fetch handlers must accept 'type' parameter
even though it's unused.

Well, it's either this or unset it in the caller, which i am considering.

I thought to leave it in in case the specified type might affect the
behavior of the handler, but so far i don't see it doing that.  So maybe
we'll want to del kw['type'] after all...