introduce new upstream.source file format in fetch_sources docstring
This is a Grand Unified upstream.source File Format; an
Ultimate Bikeshed where all formats are honored and welcomed as friends!

Hear ye! Here we:

1. Provide a mechanism to easily add checksums for our cached upstream
source files, to provide some certainty about what we are building

2. Attempt to reconcile the older, traditional, pre-meta_url format,
which accepts a single URL or upstream-cache relative-path with the
newer meta_url format coming into favor, integrating both into a
single cohesive format, and not breaking any existing syntax

3. Break up the process_meta_url monolith into more modular parts,
allowing more flexible extending of the .source 'type' system

4. Add a simple utility script to fetch sources from an upstream.source
file, with an option for whether or not to save the rpm spec file for
git sources

We begin slowly, with just the docstring to introduce the concept.

What follows will be a massively-squashed rebase of the original work,
the new bits narrated in piecemeal commits.

...

For the very curious, the original history context can be found here:

https://github.com/edquist/osg-build/compare/269cc930...c9e49a0c