introduce git_archive_remote_ref: shallow fetch + archive
this is an alternative approach to doing a full git-clone, which
features:

1. shallow fetch

Because apparently some old versions of git that we still support
do not work correctly given a tag (as opposed to a head) as the arg
for --single-branch, doing a shallow clone directly (with --depth=1)
is not an option.  But it seems we can still fetch a tag this way.

This avoids having to fetch a huge amount of extra objects from the
git server for large repos, and it does not sacrifice any of the
consistency checking of git for the result archive.

For example: a full clone of the htcondor git repo, not including any
of the checked out files, is 242M, vs only 14M for a shallow clone.

2. bare repo

There is no real need to make a mess of the local filesystem with the
unpacked copies of all the source files.  Git will make the archive
without paying any attention to them anyway, and if individual files
are needed to be copied out (eg, the rpm spec), git can do that
directly without needing a working copy.

3. no temporary uncompressed archive

Despite the existence of the --format=tar.gz option on modern (and even
old) versions of git for years now, we apparently still support the
SL6 version of git, which does not support this output format.

This means we have to run gzip ourselves, but writing an uncompressed
version to disk first is a bit wasteful.

So we avoid this by pipelining git-archive | gzip, making sure to
check the exit status of both.

...

Note that the git commands as specified in this function will set up a
git repo in GIT_DIR if defined or else the current directory.  When we
actually call this function, we will set up a temporary GIT_DIR for it.