git_archive_remote_ref: add parameter to check commit hash
Note the condition here:

+    if hash or not ops.nocheck:
+        check_git_hash(...)

means that the check runs (1) if hash is specified at all, or (2) if
the --nocheck option is not specified.

The very slight (but i think good) semantic difference from the
previous behavior is that when the --nocheck option is specified, the
hash= parameter may be omitted entirely, rather than providing a bogus
hash=0 which will always generate a warning.

Official builds don't run with --nocheck, so they will still require
a hash to be provided.  But for development work with scratch builds,
if you use --nocheck, you can just omit the hash parameter.