add fetch handler for 'git' source type
note that we've spiced things up with a few extra options compared
to what was previously supported by process_meta_url;

Namely:

1. there is an optional "spec" parameter to give a path for the rpm spec
file other than the default rpm/{name}.spec

2. there is now an optional "prefix" parameter, in addition to the
existing "tarball" option

This allows, for instance, specifying a tarball named
"vo-client-85-osg.tar.gz" with an archive prefix dir named
"vo-client-85".  (The point being, it is now allowed for the two not to
match.)

If left unspecified, as before, the archive prefix will be the same as
tarball, with the .tar.gz extension stripped, or
{name}-{tarball_version} if no tarball is specified

If prefix is specified, tarball will default to {prefix}.tar.gz

3. tarball_version will default to the basename of tag, with a leading
'v' stripped.  This means tag=wip/v1.2.3 -> tarball_version=1.2.3,
which affects the default prefix and tarball name accordingly.

...

Other notes:

- the git_archive_remote_ref() call is passed inside a lambda, so that
it is run after the temp git dir is set up

- the _nvl helper provides a default argument when an arg is None:

sometimes you can say "arg or default", and that's just fine, but nvl
will distinguish between "spec=" (an explicit empty value, which
specifies that the rpm spec from the git repo is unwanted, even if it
exists) and spec unspecified, which will default to rpm/{name}.spec