introduce get_auto_source_type
as promised, if the type= field is not provided, the first unnamed
arg will be used to determine the implicit type.  don't use this
feature unless that's what you want and you know how to use it.

caution: for type=git/github, you must use the trailing .git
extension for the auto-type to work correctly, otherwise the arg will
be interpreted as an ordinary uri or cached path