drop extra raise
at some wip point i think fancy_source_error would raise an Error if
it understood what went wrong, otherwise it would fall through, which
this extra raise was supposed to be there for, to re-raise the original
TypeError.  but it seems we're not doing that in fancy_source_error
anymore.