disallow implicit git/github types
mat got all grumpy about not wanting to confuse the users with things
happening "automatically" -- even if they are well defined.  Apparently
it is related somehow to his fondness for explicit lyrics.

anyway, i have trouble relating, as the worst case is the user gets an
error for making a mistake; but anyway i'm willing to play along.